Fixed bug threadinterrupt function that prevented script thread from being stopped with specific versions of python.

+++ fr0stlib/threadinterrupt.py
@@ -30,7 +30,8 @@ def interrupt(thread, exctype=ThreadInterrupt):
     """Raises an exception in a thread"""
     if not inspect.isclass(exctype):
         raise TypeError("Only types can be raised (not instances)")
-    res = ctypes.pythonapi.PyThreadState_SetAsyncExc(thread.ident,ctypes.py_object(exctype))
+    res = ctypes.pythonapi.PyThreadState_SetAsyncExc(ctypes.c_long(thread.ident),
+                                                     ctypes.py_object(exctype))
     # res 0 means the thread doesn't exist (or has alreaday finished)
     # res 1 is expected exit status.
     if res > 1:
